+++ pkg/model/currency.go
@@ -1,7 +1,7 @@
 package model
 
 type Currency struct {
-	Id   int64  `db:"id"`
+	Id   int    `db:"id"`
 	Code string `db:"code"`
 	Name string `db:"name"`
 }
