+++ pkg/model/operation.go
@@ -15,8 +15,8 @@ const (
 )
 
 type Operation struct {
-	Id          int64     `db:"id"`
-	IdAccount   int64     `db:"id_account"`
+	Id          int       `db:"id"`
+	IdAccount   int       `db:"id_account"`
 	Time        time.Time `db:"time"`
 	Sum         float32   `db:"sum"`
 	Category    Category  `db:"category"`
