+++ pkg/model/user.go
@@ -1,7 +1,7 @@
 package model
 
 type User struct {
-	Id                int64  `db:"id"`
+	Id                int    `db:"id"`
 	Name              string `db:"name"`
 	DefaultCurrencyId int    `db:"id_default_currency"`
 	Accounts          []Account
