+++ pkg/store/teststore/userrepository.go
@@ -14,7 +14,7 @@ type UserRepository struct {
 // Create ...
 func (r *UserRepository) Create(u *model.User) error {
 
-	u.Id = int64(len(r.users) + 1)
+	u.Id = int(len(r.users) + 1)
 	r.users[int(u.Id)] = u
 
 	return nil
